Added feature of survey ownership
Added `ownerId` to the survey model which stores the Firebase user id of
the user who creates the survey.
Modified the `edit-survey` component to store the uid during survey
creation.
Modified the dashboard page so that only surveys belonging to the logged-in
user are shown.
Modified the `view-survey` component accordingly.